Add square ring-compat logo
The rustdoc logo image rules have changed again and non-square images
are now stretched.

This adds a square logo for rustdoc.

Unfortunately an SVG version of this logo isn't available (I need to
recreate it as one), but it's less important in this application as the
latest rustdoc renderer adds a white stroke around the outside on dark
templates.